+++ src/main/java/com/barvegas/backend/Service/SerCliente.java
@@ -2,6 +2,8 @@ package com.barvegas.backend.Service;
 
 import com.barvegas.backend.Model.ModCliente;
 import com.barvegas.backend.Repository.RepCliente;
+import com.barvegas.backend.exception.BadRequestException;
+import com.barvegas.backend.exception.ServerErrorException;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,16 +26,16 @@ public class SerCliente {
     public ModCliente getClienteById(Long id){
         Optional<ModCliente> optionalMembro_canditatoModel =
                                         repCliente.findById(id);
-
         if(!optionalMembro_canditatoModel.isPresent())
-            throw new MethodNotFoundException("Produto não encontrado...");
-
+            throw new BadRequestException("Produto não encontrado...");
         return optionalMembro_canditatoModel.get();
     }
 
     //Alterar Membro/Candidato (Todos os campos)
     public ModCliente updateClienteById(Long id , ModCliente upCliente){
-
+        if(repCliente.findById(id).isEmpty()){
+            throw new BadRequestException("O cliente procurado não existe");
+        }
         upCliente.setId(id);
         return repCliente.save(upCliente);
     }
@@ -46,6 +48,9 @@ public class SerCliente {
 
     //Deletar Produto por código
     public void deleteClienteById(Long id){
+        if(repCliente.findById(id).isEmpty()){
+            throw new ServerErrorException("O cliente que deseja deletar não existe.");
+        }
         repCliente.deleteById(id);
     }
